ipv6: Add support for requesting a prefix with DHCPv6 PD
By setting the 'request_prefix' attribute to '1' dhclient will
be invoked with the -P and -N flag which will cause it to request
both an address and a prefix.

A example configuration for a interface would be:

  iface eno1 inet6 dhcp
      autoconf 0
      request_prefix 1

Or

  iface eno1 inet6 auto
      dhcp 1
      request_prefix 1

The obtained prefix can be found in the leases file. Other ways to see
the prefix is by using a hook which dhclient executes. The $new_ip6_prefix
variable will contain the IPv6 prefix including the netmask.

IPv6 Prefix Delegation can be used for multiple scenarios. For example
where a Linux machine acts as a router and requests a prefix from upstream
which it can use to provide IPv6 connectivity downstream.

Other examples are Virtualization of Containers where VMs or Containers
get native IPv6 access by a prefix the host obtained through Prefix Delegation.

See 'man dhclient' for more information.

Signed-off-by: Wido den Hollander <wido@widodh.nl>