Flush automatically assigned addresses on ifup when accept_ra=0.
There is a race condition in the kernel; by default it accepts route
advertisements, so by the time ifup runs and tries to disable accept_ra,
the interface might already have acquired an address. So first set
accept_ra to 0, then flush any address marked mngtmpaddr on the
interface, then do the rest of the configuration.