Configure VLANs on bridge interfaces.
Bridge-utils integrated with the vlan package, so that bridge interfaces
with VLANs would be handled by the vlan packge. Ifupdown can also
configure VLANs, but it had a check that prevented it from doing that
for bridge interfaces. However, bridge-utils 1.5-11 removed the
integration with the vlan package, and now relies on us to configure
VLANs.

See also bug #818849.