First stab at pattern matching interfaces.
When building the list of interfaces to act upon, ifup/ifdown will now
expand interfaces with a slash in the name differently. The final goal
is to allow the following form:

    [variable]/pattern[/options]

At the moment however, only the following form is handled:

    /pattern

When the above form is encountered, it looks at the list of interfaces
known by the kernel, and uses fnmatch() to match each of those against
the pattern. Any matches are acted upon.

It is also still possible to specify which logical interface to use for
any matching real interfaces. So for example:

    auto /eth*=eth
    iface eth inet dhcp

Will match any real interface whose name starts with eth, and uses the
"iface eth" stanza to configure that interface.

Slashes are not allowed in interface names (at least not on Linux), so
there is no need to worry about backward compatibility.