Fix Infiniband device detection
Detecting an Infiniband interface by testing for the existence of
/sys/class/net/%link%/device/infiniband is unreliable, as an
Infiniband capable port running in Ethernet mode may still have
the Infiniband directory in sysfs. When this type of port is in
Ethernet mode, ifupdown attempts and fails to create an Infiniband
partition on the Ethernet port, instead of creating a VLAN.

This patch instead only creates Infiniband partitions when an
interface's type is ARPHRD_INFINIBAND (32).