Handle interface renaming in ifupdown-hotplug.
When an interface is hotplugged, the ifupdown-hotplug script first
checks whether the interface is listed as an allow-hotplug iface.
However, it didn't take care of renaming, so the following:

    rename foo=bar
    allow-hotplug bar
    iface bar ...

Would cause it to ignore the interface called foo, even though "ifup
--allow hotplug foo" would work correctly. Use the new "ifquery --allow
hotplug --list <iface>" method to find out if foo is allowed.

When using systemd, and an interface is hotplugged, udev will call the
ifupdown-hotplug script with the current name of the interface as the
argument. The script will start ifup@currentname, which will work fine,
but the interface might subsequencly be renamed by ifup. This causes a
problem when the ifup@ service is stopped, since it tries to ifdown the
old name. So use the output of ifquery to start the service with the new
name.