Buffer SIGINT and SIGTERM.
When we get either of those signals, just set a flag that gets checked
whenever we process an interface or a command. This allows for a more
graceful termination; in particular it will leave the state files in a
more consistent state.

After SIGINT or SIGTERM has been received once, the signal handler is
reverted back to the default handler.