hotfix: don't hang on boot when dhcp is unreachable
There is a conflict between /lib/udev/ifupdown-hotplug and
/etc/init.d/networking that causes boot to hang several
minutes when interface is configured to  DHCP and doesn't
get any DHCP response.

In that case the udev trigger will try for several minutes
and keeping the interface locked, while the ifup_hotplug()
function tries to configure the interface again, causing the
bootup to stall because of waiting for the lock.

IMHO, this is function is completely wrong in case of the
udev trigger already doing the probing.

Hotfix by just returning out of the function. A clean solution
should check whether interface is already locked and skip it
in this case.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>